+++ plugins/image-trust/pkg/config/modes.go
@@ -1,0 +1,8 @@
+package config
+
+const (
+	ModeCosignKeyless = "cosign-keyless"
+	ModeCosignKey     = "cosign-key"
+
+	ModePolicyAny = "any"
+)